Fix uint32_t formatting warnings in alpha_hwr.cpp

+++ components/alpha_hwr/alpha_hwr.cpp
@@ -107,8 +107,8 @@ void AlphaHwrComponent::setup() {
     // single-pump node it means "the pump is bonded."
     if (this->reconnect_settle_ms_ > 0 && this->parent_ != nullptr &&
         esp_ble_get_bond_device_num() > 0) {
-      ESP_LOGI(TAG, "Disconnected; holding reconnect until pump reappears + %u ms",
-               this->reconnect_settle_ms_);
+      ESP_LOGI(TAG, "Disconnected; holding reconnect until pump reappears + %lu ms",
+               (unsigned long)this->reconnect_settle_ms_);
       this->parent_->set_auto_connect(false);
       this->reconnect_settling_ = true;
       this->reconnect_timer_armed_ = false;
@@ -272,8 +272,8 @@ bool AlphaHwrComponent::parse_device(
   }
   // Pump is advertising again after the disconnect — start the settle timer once.
   this->reconnect_timer_armed_ = true;
-  ESP_LOGI(TAG, "Pump reappeared; holding reconnect %u ms to let it settle",
-           this->reconnect_settle_ms_);
+  ESP_LOGI(TAG, "Pump reappeared; holding reconnect %lu ms to let it settle",
+           (unsigned long)this->reconnect_settle_ms_);
   this->set_timeout("reconnect_settle", this->reconnect_settle_ms_, [this]() {
     ESP_LOGI(TAG, "Reconnect settle window elapsed; allowing reconnect");
     this->reconnect_settling_ = false;
@@ -710,8 +710,8 @@ void AlphaHwrComponent::read_statistics() {
   device_info_service_.read_statistics_async(
       [this](bool success, uint32_t start_count, float operating_hours) {
         if (success) {
-          ESP_LOGI(TAG, "Statistics read successful: %u starts, %.1f hours",
-                   start_count, operating_hours);
+          ESP_LOGI(TAG, "Statistics read successful: %lu starts, %.1f hours",
+                   (unsigned long)start_count, operating_hours);
           if (this->start_count_sensor_) {
             this->start_count_sensor_->publish_state(start_count);
           }
